Improve certificate validation and move to its own file
